order and orderresults endpints

+++ backend/db/get_user_by_id.sql
@@ -1,2 +1,2 @@
 select * from users
-where userid = ${userid}
+where userid::integer = ${userid}
